+++ backend/common/settings/structs.go
@@ -13,31 +13,30 @@ const (
 )
 
 type Settings struct {
-	Server       Server         `json:"server"`
-	Auth         Auth           `json:"auth"`
-	Frontend     Frontend       `json:"frontend"`
-	Users        []UserDefaults `json:"users,omitempty"`
-	UserDefaults UserDefaults   `json:"userDefaults"`
-	Integrations Integrations   `json:"integrations"`
+	Server       Server       `json:"server"`
+	Auth         Auth         `json:"auth"`
+	Frontend     Frontend     `json:"frontend"`
+	UserDefaults UserDefaults `json:"userDefaults"`
+	Integrations Integrations `json:"integrations"`
 }
 
 type Server struct {
-	NumImageProcessors           int         `json:"numImageProcessors"`
-	Socket                       string      `json:"socket"`
-	TLSKey                       string      `json:"tlsKey"`
-	TLSCert                      string      `json:"tlsCert"`
-	DisablePreviews              bool        `json:"disablePreview"`
-	ResizePreviews               bool        `json:"resizePreview"`
-	DisableTypeDetectionByHeader bool        `json:"disableTypeDetectionByHeader"`
-	Port                         int         `json:"port"`
-	BaseURL                      string      `json:"baseURL"`
+	NumImageProcessors           int         `json:"numImageProcessors"`           // number of concurrent image processing jobs
+	Socket                       string      `json:"socket"`                       // socket to listen on
+	TLSKey                       string      `json:"tlsKey"`                       // path to TLS key
+	TLSCert                      string      `json:"tlsCert"`                      // path to TLS cert
+	DisablePreviews              bool        `json:"disablePreviews"`              // disable all previews thumbnails, simple icons will be used
+	DisableResize                bool        `json:"disablePreviewResize"`         // disable resizing of previews for faster loading over slow connections
+	DisableTypeDetectionByHeader bool        `json:"disableTypeDetectionByHeader"` // disable type detection by header, useful if filesystem is slow.
+	Port                         int         `json:"port"`                         // port to listen on
+	BaseURL                      string      `json:"baseURL"`                      // base URL for the server, the subpath that the server is running on.
 	Logging                      []LogConfig `json:"logging"`
-	Database                     string      `json:"database"`
+	Database                     string      `json:"database"` // path to the database file
 	Sources                      []Source    `json:"sources" validate:"required,dive"`
-	ExternalUrl                  string      `json:"externalUrl"`
-	InternalUrl                  string      `json:"internalUrl"` // used by integrations
-	CacheDir                     string      `json:"cacheDir"`
-	MaxArchiveSizeGB             int64       `json:"maxArchiveSize"`
+	ExternalUrl                  string      `json:"externalUrl"`    // used by share links if set
+	InternalUrl                  string      `json:"internalUrl"`    // used by integrations if set, this is the url that an integration service will use to communicate with filebrowser
+	CacheDir                     string      `json:"cacheDir"`       // path to the cache directory, used for thumbnails and other cached files
+	MaxArchiveSizeGB             int64       `json:"maxArchiveSize"` // max pre-archive combined size of files/folder that are allowed to be archived (in GB)
 	// not exposed to config
 	SourceMap     map[string]Source `json:"-" validate:"omitempty"` // uses realpath as key
 	NameToSource  map[string]Source `json:"-" validate:"omitempty"` // uses name as key
@@ -52,20 +51,20 @@ type Integrations struct {
 // onlyoffice secret is stored in the local.json file
 // docker exec <containerID> /var/www/onlyoffice/documentserver/npm/json -f /etc/onlyoffice/documentserver/local.json 'services.CoAuthoring.secret.session.string'
 type OnlyOffice struct {
-	Url    string `json:"url" validate:"required"`
+	Url    string `json:"url" validate:"required"` // URL to the OnlyOffice Document Server
 	Secret string `json:"secret" validate:"required"`
 }
 
 type Media struct {
-	FfmpegPath string `json:"ffmpegPath"`
+	FfmpegPath string `json:"ffmpegPath"` // path to ffmpeg binary, including the binary name (eg. /usr/bin/ffmpeg)
 }
 
 type LogConfig struct {
-	Levels    string `json:"levels"`
-	ApiLevels string `json:"apiLevels"`
-	Output    string `json:"output"`
-	NoColors  bool   `json:"noColors"`
-	Json      bool   `json:"json"`
+	Levels    string `json:"levels"`    // separated list of log levels to enable. (eg. "info|warning|error|debug")
+	ApiLevels string `json:"apiLevels"` // separated list of log levels to enable for the API. (eg. "info|warning|error")
+	Output    string `json:"output"`    // output location. (eg. "stdout" or "path/to/file.log")
+	NoColors  bool   `json:"noColors"`  // disable colors in the output
+	Json      bool   `json:"json"`      // output in json format, currently not supported
 }
 
 type Source struct {
@@ -75,36 +74,36 @@ type Source struct {
 }
 
 type SourceConfig struct {
-	IndexingInterval      uint32      `json:"indexingInterval"`
-	Disabled              bool        `json:"disabled"`
-	MaxWatchers           int         `json:"maxWatchers"`
-	NeverWatch            []string    `json:"neverWatchPaths"`
-	IgnoreHidden          bool        `json:"ignoreHidden"`
-	IgnoreZeroSizeFolders bool        `json:"ignoreZeroSizeFolders"`
-	Exclude               IndexFilter `json:"exclude"`
-	Include               IndexFilter `json:"include"`
-	DefaultUserScope      string      `json:"defaultUserScope"` // default "" should match folders under path
-	DefaultEnabled        bool        `json:"defaultEnabled"`
-	CreateUserDir         bool        `json:"createUserDir"`
+	IndexingInterval      uint32      `json:"indexingInterval"`      // optional manual overide interval in seconds to re-index the source
+	Disabled              bool        `json:"disabled"`              // disable the indexing of this source
+	MaxWatchers           int         `json:"maxWatchers"`           // number of concurrent watchers to use for this source, currently not supported
+	NeverWatch            []string    `json:"neverWatchPaths"`       // paths to never watch, relative to the source path (eg. "/folder/file.txt")
+	IgnoreHidden          bool        `json:"ignoreHidden"`          // ignore hidden files and folders.
+	IgnoreZeroSizeFolders bool        `json:"ignoreZeroSizeFolders"` // ignore folders with 0 size
+	Exclude               IndexFilter `json:"exclude"`               // exclude files and folders from indexing, if include is not set
+	Include               IndexFilter `json:"include"`               // include files and folders from indexing, if exclude is not set
+	DefaultUserScope      string      `json:"defaultUserScope"`      // default "/" should match folders under path
+	DefaultEnabled        bool        `json:"defaultEnabled"`        // should be added as a default source for new users?
+	CreateUserDir         bool        `json:"createUserDir"`         // create a user directory for each user
 }
 
 type IndexFilter struct {
-	Files        []string `json:"files"`
-	Folders      []string `json:"folders"`
-	FileEndsWith []string `json:"fileEndsWith"`
+	Files        []string `json:"files"`        // array of file names to include/exclude
+	Folders      []string `json:"folders"`      // array of folder names to include/exclude
+	FileEndsWith []string `json:"fileEndsWith"` // array of file names to include/exclude (eg "a.jpg")
 }
 
 type Frontend struct {
-	Name                  string         `json:"name"`
-	DisableDefaultLinks   bool           `json:"disableDefaultLinks"`
-	DisableUsedPercentage bool           `json:"disableUsedPercentage"`
+	Name                  string         `json:"name"`                  // display name
+	DisableDefaultLinks   bool           `json:"disableDefaultLinks"`   // disable default links in the sidebar
+	DisableUsedPercentage bool           `json:"disableUsedPercentage"` // disable used percentage for the sources in the sidebar
 	ExternalLinks         []ExternalLink `json:"externalLinks"`
 }
 
 type ExternalLink struct {
-	Text  string `json:"text" validate:"required"`
-	Title string `json:"title"`
-	Url   string `json:"url" validate:"required"`
+	Text  string `json:"text" validate:"required"` // the text to display on the link
+	Title string `json:"title"`                    // the title to display on hover
+	Url   string `json:"url" validate:"required"`  // the url to link to
 }
 
 // UserDefaults is a type that holds the default values