+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file. For commit 
   - Updated placeholder for scopes https://github.com/gtsteffaniak/filebrowser/issues/475
   - When user's API permissions are removed, any api keys the user had will be revoked.
   - `server.enableThumbnails` moved to `server.disablePreviews` defaulting to false.
+  - `server.resizePreview` moved to `server.resizePreviews` (with an "s" at the end)
 
   **Bug Fixes**:
   - Nil pointer error when source media is disconnected while running.